rename folder to just survey
